Add Content type definition (#6381)
* rename the definition file of another `content-type` library

* add the definition of `content-type` library
