+++ content-type/content-type-tests.ts
@@ -1,47 +1,21 @@
 /// <reference path="content-type.d.ts" />
+/// <reference path="../express/express.d.ts" />
 
-import MediaType = require('content-type');
+import contentType = require('content-type');
+import express = require('express');
 
-// https://github.com/deoxxa/content-type/blob/master/README.md
-function new_test(): void {
-    var p = new MediaType('text/html;level=1;q=0.5');
-    p.q === 0.5;
-    p.params.level === "1";
 
-    var q = new MediaType('application/json', { profile: 'http://example.com/schema.json' });
-    q.type === "application/json";
-    q.params.profile === "http://example.com/schema.json";
+var obj = contentType.parse('image/svg+xml; charset=utf-8');
 
-    q.q = 1;
-    q.toString() === 'application/json;q=1;profile="http://example.com/schema.json"';
-}
+console.log(obj.type);  // => 'image/svg+xml'
+console.log(obj.parameters.charset);  // => 'utf-8'
 
-function mediaCmp_test(): void {
-    MediaType.mediaCmp(MediaType.parseMedia('text/html'), MediaType.parseMedia('text/html')) === 0;
-    MediaType.mediaCmp(MediaType.parseMedia('*/*'), MediaType.parseMedia('text/html')) === 1;
-    MediaType.mediaCmp(MediaType.parseMedia('text/html;level=1'), MediaType.parseMedia('text/html')) === -1;
-    MediaType.mediaCmp(MediaType.parseMedia('application/json;profile="v1.json"'), MediaType.parseMedia('application/json;profile="v2.json"')) === null;
-}
 
-// https://github.com/deoxxa/content-type/blob/master/example.js
-function example(): void {
-    var representations = [
-        'application/json',
-        'text/html',
-        'application/json;profile="schema.json"',
-        'application/json;profile="different.json"',
-    ];
+var req: express.Request;
+obj = contentType.parse(req);
 
-    var accept = [
-        'text/html;q=0.50',
-        '*/*;q=0.01',
-        'application/json;profile=different.json',
-        'application/json;profile="a,b;c.json?d=1;f=2";q=0.2',
-    ];
+var res: express.Response;
+obj = contentType.parse(res);
 
-    console.log('Formats:\n\t' + representations.map(MediaType.parseMedia).join('\n\t'));
+var str: string = contentType.format({type: 'image/svg+xml'});
 
-    console.log('Accept:\n\t' + accept.map(MediaType.parseMedia).join('\n\t'));
-
-    console.log('Selected:', (MediaType.select(representations.map(MediaType.parseMedia), accept.map(MediaType.parseMedia)) || 'None').toString());
-}
